ConderenceId getter no longer implicitly casts to long

+++ src/main/java/com/cms/entity/Conference.java
@@ -55,7 +55,7 @@ public class Conference {
 		this.accomodationInfo = accomodationInfo;
 	}
 
-	public long getConferenceID() {
+	public Integer getConferenceID() {
 		return conferenceID;
 	}
